Completed Wave 1, really, by separating View and Model from app.js and put them in correct view and model files.
